Code refactorisation

+++ OpenGLCourseAPP/OpenGLCourseAPP/GLWindow.h
@@ -13,8 +13,8 @@ public:
 
 	int Initialise();
 
-	GLfloat getBufferWidth() { return bufferWidth; }
-	GLfloat getBufferHeight() { return bufferHeight; }
+	GLint getBufferWidth() { return bufferWidth; }
+	GLint getBufferHeight() { return bufferHeight; }
 
 	bool getShouldClose() { return glfwWindowShouldClose(mainWindow); }
